backend/requirements.txt
uvicorn
sqlmodel
google-auth
python-dotenv
psycopg2-binary
python-jose[cryptography]
requests
